pyproject.toml
[tool.poetry]
name = "pro-vision"
version = "0.1.2"
description = "Mapping public service access points, and their accessibility across Chicago"
authors = ["Setu"]
license = "N"
readme = "README.md"
packages = [{include = "pro_vision"}]

[tool.poetry.dependencies]
python = "3.8.10"
geopandas = "^0.12.2"
pandas = "^1.5.3"
traveltimepy = "^3.0.3"
requests = "^2.28.2"
googlemaps = "^4.10.0"
DateTime = "^5.0"
dash = "^2.8.1"
dash-bootstrap-components = "^1.4.0"
plotly-express = "^0.4.1"
openpyxl = "^3.1.1"
geojson = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
